+++ tests/display_text.sql
@@ -1,0 +1,1 @@
+select 'html' as component, $html as html;