+++ tests/sql_test_files/it_works_run_sql_from_database.sql
@@ -1,0 +1,7 @@
+select 'dynamic' as component,
+    sqlpage.run_sql('tests/display_text.sql', CONCAT('{"html": "', html,'"}')) as properties
+from (
+    select 'It ' as html
+    union all
+    select 'works !'
+);